follow style guide, typo fixes
